About variations on algortihm seed
A new draft has been added. There are problems when too few data is used. Now in positive class there are 100 resources, so in a 10 fold (CV), on validation time there is 10 resources. Thus, in a imbalance case (1 vs 100), if no TP are found, then NaN appeared when metrics are calculated.